Update script and instructions
